\docType{package}
\name{RcppParallel-package}
\alias{RcppParallel}
\alias{RcppParallel-package}
\title{Parallel programming tools for Rcpp}
\description{
  High level functions for doing parallel programming with Rcpp.
  For example, the parallelFor function can be used to convert the work of 
  a standard serial "for" loop into a parallel one and the parallelReduce 
  function can be used for accumulating aggregate or other values.
}
\author{
  JJ Allaire \email{jj@rstudio.com}
}
\keyword{package}

