\docType{package}
\name{tbb-package}
\alias{tbb}
\alias{tbb-package}
\title{Interface to Intel TBB (Threading Building Blocks)}
\description{
   Interface to Intel TBB (Threading Building Blocks), a C++ library
   for parallel programming. It uses C++ templates to provide powerful parallel 
   functionality that works with common programming patterns. For example, 
   the parallel_for construct can be used to convert the work of a standard 
   serial "for"" loop into a parallel one and the parallel_reduce construct
   can be used for accumulating aggregate or other values.
}
\author{
  JJ Allaire \email{jj@rstudio.com}
}
\keyword{package}

