% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbb.R
\name{tbbLibPath}
\alias{tbbLibPath}
\title{Get the Path to a TBB Library}
\usage{
tbbLibPath(name = "tbb")
}
\arguments{
\item{name}{The name of the library. Normally, this is one of
`tbb`, `tbbmalloc`, or `tbbmalloc_proxy`.}
}
\description{
Retrieve the path to a TBB library. This can be useful for \R packages
using RcppParallel that wish to use, or re-use, the version of TBB that
RcppParallel has been configured to use.
}
