% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbb.R
\name{tbbLibraryPath}
\alias{tbbLibraryPath}
\title{Get the Path to a TBB Library}
\usage{
tbbLibraryPath(name = NULL)
}
\arguments{
\item{name}{The name of the TBB library to be resolved. Normally, this is one of
\code{tbb}, \code{tbbmalloc}, or \code{tbbmalloc_proxy}. When \code{NULL}, the library
path containing the TBB libraries is returned instead.}
}
\description{
Retrieve the path to a TBB library. This can be useful for \R packages
using RcppParallel that wish to use, or re-use, the version of TBB that
RcppParallel has been configured to use.
}
